Implement vendor model with package management and installation logic
