+++ internal/exec/vendor_utils.go
@@ -1,20 +1,13 @@
 package exec
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"path"
 	"path/filepath"
-	"strconv"
 	"strings"
-	"time"
 
-	"github.com/charmbracelet/bubbles/progress"
-	"github.com/charmbracelet/bubbles/spinner"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
-	"github.com/hashicorp/go-getter"
 	cp "github.com/otiai10/copy"
 	"github.com/samber/lo"
 	"github.com/spf13/cobra"
@@ -23,232 +16,6 @@ import (
 	"github.com/cloudposse/atmos/pkg/schema"
 	u "github.com/cloudposse/atmos/pkg/utils"
 )
-
-type pkgType int
-
-const (
-	pkgTypeRemote pkgType = iota
-	pkgTypeOci
-	pkgTypeLocal
-)
-
-type pkg struct {
-	uri               string
-	name              string
-	targetPath        string
-	cliConfig         schema.CliConfiguration
-	s                 schema.AtmosVendorSource
-	sourceIsLocalFile bool
-	pkgType           pkgType
-	version           string
-}
-type model struct {
-	packages  []pkg
-	index     int
-	width     int
-	height    int
-	spinner   spinner.Model
-	progress  progress.Model
-	done      bool
-	dryRun    bool
-	failedPkg int
-}
-
-var (
-	currentPkgNameStyle = lipgloss.NewStyle().Foreground(lipgloss.Color("211"))
-	doneStyle           = lipgloss.NewStyle().Margin(1, 2)
-	checkMark           = lipgloss.NewStyle().Foreground(lipgloss.Color("42")).SetString("✓")
-	xMark               = lipgloss.NewStyle().Foreground(lipgloss.Color("9")).SetString("x")
-)
-
-func newModel(pkg []pkg, dryRun bool) (model, error) {
-	p := progress.New(
-		progress.WithDefaultGradient(),
-		progress.WithWidth(40),
-		progress.WithoutPercentage(),
-	)
-	s := spinner.New()
-	s.Style = lipgloss.NewStyle().Foreground(lipgloss.Color("63"))
-
-	return model{
-		packages: pkg,
-		spinner:  s,
-		progress: p,
-		dryRun:   dryRun,
-	}, nil
-}
-
-func (m model) Init() tea.Cmd {
-	// Start downloading with the `uri`, package name, and `tempDir` directly from the model
-	return tea.Batch(downloadAndInstall(m.packages[0], m.dryRun), m.spinner.Tick)
-}
-func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-	case tea.WindowSizeMsg:
-		m.width, m.height = msg.Width, msg.Height
-	case tea.KeyMsg:
-		switch msg.String() {
-		case "ctrl+c", "esc", "q":
-			return m, tea.Quit
-		}
-	case installedPkgMsg:
-
-		pkg := m.packages[m.index]
-		mark := checkMark
-
-		if msg.err != nil {
-			u.LogDebug(pkg.cliConfig, fmt.Sprintf("Failed to pull component %s error %s", pkg.name, msg.err))
-			mark = xMark
-			m.failedPkg++
-		}
-		if m.index >= len(m.packages)-1 {
-			// Everything's been installed. We're done!
-			m.done = true
-			return m, tea.Sequence(
-				tea.Printf("%s %s %s", mark, pkg.name, pkg.version),
-				tea.Quit,
-			)
-		}
-
-		// Update progress bar
-		m.index++
-		progressCmd := m.progress.SetPercent(float64(m.index) / float64(len(m.packages)))
-		return m, tea.Batch(
-			progressCmd,
-			tea.Printf("%s %s %s", mark, pkg.name, pkg.version), // print success message above our program
-			downloadAndInstall(m.packages[m.index], m.dryRun),   // download the next package
-		)
-	case spinner.TickMsg:
-		var cmd tea.Cmd
-		m.spinner, cmd = m.spinner.Update(msg)
-		return m, cmd
-	case progress.FrameMsg:
-		newModel, cmd := m.progress.Update(msg)
-		if newModel, ok := newModel.(progress.Model); ok {
-			m.progress = newModel
-		}
-		return m, cmd
-	}
-	return m, nil
-}
-
-func (m model) View() string {
-	n := len(m.packages)
-	w := lipgloss.Width(fmt.Sprintf("%d", n))
-	if m.done {
-		if m.dryRun {
-			return doneStyle.Render("Done! Dry run completed. No components installed.\n")
-		}
-		if m.failedPkg > 0 {
-			return doneStyle.Render(fmt.Sprintf("Done! Installed %d components. %d components failed to install.\n", n-m.failedPkg, m.failedPkg))
-		}
-		return doneStyle.Render(fmt.Sprintf("Done! Installed %d components.\n", n))
-	}
-
-	pkgCount := fmt.Sprintf(" %*d/%*d", w, m.index, w, n)
-	spin := m.spinner.View() + " "
-	prog := m.progress.View()
-	cellsAvail := max(0, m.width-lipgloss.Width(spin+prog+pkgCount))
-
-	pkgName := currentPkgNameStyle.Render(m.packages[m.index].name)
-
-	info := lipgloss.NewStyle().MaxWidth(cellsAvail).Render("Pulling " + pkgName)
-
-	cellsRemaining := max(0, m.width-lipgloss.Width(spin+info+prog+pkgCount))
-	gap := strings.Repeat(" ", cellsRemaining)
-
-	return spin + info + gap + prog + pkgCount
-}
-
-type installedPkgMsg struct {
-	err  error
-	name string
-}
-
-func downloadAndInstall(p pkg, dryRun bool) tea.Cmd {
-	return func() tea.Msg {
-		if dryRun {
-			// Simulate the action
-			time.Sleep(1 * time.Second)
-			return installedPkgMsg{
-				err:  nil,
-				name: p.name,
-			}
-		}
-		// Create temp directory
-		tempDir, err := os.MkdirTemp("", strconv.FormatInt(time.Now().Unix(), 10))
-		if err != nil {
-			return err
-		}
-		defer removeTempDir(p.cliConfig, tempDir)
-
-		switch p.pkgType {
-		case pkgTypeRemote:
-			// Use go-getter to download remote packages
-			client := &getter.Client{
-				Ctx:  context.Background(),
-				Dst:  tempDir,
-				Src:  p.uri,
-				Mode: getter.ClientModeAny,
-			}
-			if err := client.Get(); err != nil {
-				return installedPkgMsg{
-					err:  err,
-					name: p.name,
-				}
-			}
-
-		case pkgTypeOci:
-			// Process OCI images
-			if err := processOciImage(p.cliConfig, p.uri, tempDir); err != nil {
-				return installedPkgMsg{
-					err:  err,
-					name: p.name,
-				}
-			}
-
-		case pkgTypeLocal:
-			// Copy from local file system
-			copyOptions := cp.Options{
-				PreserveTimes: false,
-				PreserveOwner: false,
-				OnSymlink:     func(src string) cp.SymlinkAction { return cp.Deep },
-			}
-			if p.sourceIsLocalFile {
-				tempDir = path.Join(tempDir, filepath.Base(p.uri))
-			}
-			if err := cp.Copy(p.uri, tempDir, copyOptions); err != nil {
-				return installedPkgMsg{
-					err:  err,
-					name: p.name,
-				}
-			}
-		default:
-			return installedPkgMsg{
-				err:  fmt.Errorf("unknown package type"),
-				name: p.name,
-			}
-
-		}
-		if err := copyToTarget(p.cliConfig, tempDir, p.targetPath, p.s, p.sourceIsLocalFile, p.uri); err != nil {
-			return installedPkgMsg{
-				err:  err,
-				name: p.name,
-			}
-		}
-		return installedPkgMsg{
-			err:  nil,
-			name: p.name,
-		}
-	}
-}
-
-func max(a, b int) int {
-	if a > b {
-		return a
-	}
-	return b
-}
 
 // ExecuteVendorPullCommand executes `atmos vendor` commands
 func ExecuteVendorPullCommand(cmd *cobra.Command, args []string) error {
@@ -529,11 +296,10 @@ func ExecuteAtmosVendorInternal(
 				uri:               uri,
 				name:              pkgName,
 				targetPath:        targetPath,
-				cliConfig:         cliConfig,
-				s:                 s,
 				sourceIsLocalFile: sourceIsLocalFile,
 				pkgType:           pType,
 				version:           s.Version,
+				s:                 s,
 			}
 
 			packages = append(packages, p)
@@ -544,7 +310,7 @@ func ExecuteAtmosVendorInternal(
 
 	// Run TUI to process packages
 	if len(packages) > 0 {
-		model, err := newModel(packages, dryRun)
+		model, err := newModel(packages, dryRun, cliConfig)
 		if err != nil {
 			return fmt.Errorf("error initializing model: %v", err)
 		}
